requirements.txt
discord.py
aiohttp
async-timeout
attrs
chardet
datetime
idna
multidict
numpy
Pillow
psutil
pyfiglet
python-dotenv
requests
typing-extensions
yarl
